refactor: move existing scrapers to git package
