+++ scrapers/src/main/java/nl/esciencecenter/rsd/scraper/git/PostgrestSIR.java
@@ -1,9 +1,10 @@
-package nl.esciencecenter.rsd.scraper;
+package nl.esciencecenter.rsd.scraper.git;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
+import nl.esciencecenter.rsd.scraper.Utils;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
