+++ scrapers/src/main/java/nl/esciencecenter/rsd/scraper/git/SoftwareInfo.java
@@ -1,4 +1,4 @@
-package nl.esciencecenter.rsd.scraper;
+package nl.esciencecenter.rsd.scraper.git;
 
 public interface SoftwareInfo {
