+++ scrapers/src/test/java/nl/esciencecenter/rsd/scraper/git/GitLabSITest.java
@@ -1,5 +1,6 @@
-package nl.esciencecenter.rsd.scraper;
+package nl.esciencecenter.rsd.scraper.git;
 
+import nl.esciencecenter.rsd.scraper.git.GitLabSI;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
